+++ test/Spec.hs
@@ -1,12 +1,12 @@
 import CoreTests (coreTests)
-import LambdaTests (lambdaTests)
 import ParserTests (parserTests)
+import TaoLangTests (taoLangTests)
 import TaoTests (taoTests)
 import Test.Hspec (hspec)
 
 main :: IO ()
 main = hspec $ do
   parserTests
-  lambdaTests
   coreTests
   taoTests
+  taoLangTests
